% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{backpropagation}
\alias{backpropagation}
\title{Backpropagation}
\usage{
backpropagation(X, Y)
}
\description{
Backpropagation
}
\seealso{
Other backpropagation: \code{\link{backpropagation_data}}

Other training: \code{\link{deeplearning}},
  \code{\link{gradientdescent}},
  \code{\link{recurrentneuralnetwork}}
}
