% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{backpropagation_data}
\alias{backpropagation_data}
\title{Generate Backpropagation Data}
\usage{
backpropagation_data(n = 50000)
}
\description{
Generate Backpropagation Data
}
\seealso{
Other backpropagation: \code{\link{backpropagation_training}}

Other data generators: \code{\link{deeplearning_data}},
  \code{\link{gradientdescent_data}},
  \code{\link{recurrentneuralnetwork_data}}
}
