% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{backpropagation_training}
\alias{backpropagation_training}
\title{Backpropagation Training}
\usage{
backpropagation_training(X, Y)
}
\description{
Backpropagation Training
}
\seealso{
Other backpropagation: \code{\link{backpropagation_data}}

Other training: \code{\link{deeplearning_training}},
  \code{\link{gradientdescent_training}},
  \code{\link{recurrentneuralnetwork_training}}
}
