% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{deeplearning}
\alias{deeplearning}
\alias{deeplearning_training}
\title{Deep Learning Training}
\usage{
deeplearning_training(X, Y, hidden_dim = 4)
}
\description{
Deep Learning Training
}
\seealso{
Other training: \code{\link{backpropagation}},
  \code{\link{gradientdescent}},
  \code{\link{recurrentneuralnetwork}}
}
