% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{deeplearning_data}
\alias{deeplearning_data}
\title{Generate Deep Learning Data}
\usage{
deeplearning_data(n = 50000)
}
\description{
Generate Deep Learning Data
}
\seealso{
Other data generators: \code{\link{backpropagation_data}},
  \code{\link{gradientdescent_data}},
  \code{\link{recurrentneuralnetwork_data}}
}
