% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{deeplearning_training}
\alias{deeplearning_training}
\title{Deep Learning Training}
\usage{
deeplearning_training(X, Y, hidden_dim = 4)
}
\description{
Deep Learning Training
}
\seealso{
Other deep learning: \code{\link{deeplearning_data}}

Other training: \code{\link{backpropagation_training}},
  \code{\link{gradientdescent_training}},
  \code{\link{recurrentneuralnetwork_training}}
}
