% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{gradientdescent}
\alias{gradientdescent}
\alias{gradientdescent_training}
\title{Gradient Descent Training}
\usage{
gradientdescent_training(X, Y, hidden_dim = 4, alpha = 0.1)
}
\description{
Gradient Descent Training
}
\seealso{
Other training: \code{\link{backpropagation}},
  \code{\link{deeplearning}},
  \code{\link{recurrentneuralnetwork}}
}
