% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\docType{data}
\name{gradientdescent_data}
\alias{gradientdescent_data}
\title{Generate Gradient Descent Data}
\format{An object of class \code{matrix} with 6 rows and 4 columns.}
\usage{
gradientdescent_data
}
\description{
Generate Gradient Descent Data
}
\seealso{
Other data generators: \code{\link{backpropagation_data}},
  \code{\link{deeplearning_data}},
  \code{\link{recurrentneuralnetwork_data}}
}
\keyword{datasets}
