% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{gradientdescent_training}
\alias{gradientdescent_training}
\title{Gradient Descent Training}
\usage{
gradientdescent_training(X, Y, hidden_dim = 4, alpha = 0.1)
}
\description{
Gradient Descent Training
}
\seealso{
Other gradient descent: \code{\link{gradientdescent_data}}

Other training: \code{\link{backpropagation_training}},
  \code{\link{deeplearning_training}},
  \code{\link{recurrentneuralnetwork_training}}
}
