% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{recurrentneuralnetwork}
\alias{recurrentneuralnetwork}
\title{Recurrent Neural Network Training}
\usage{
recurrentneuralnetwork(X1, X2, Y, hidden_dim = 10, alpha = 0.1)
}
\description{
Recurrent Neural Network Training
}
\seealso{
Other training: \code{\link{backpropagation}},
  \code{\link{deeplearning}}, \code{\link{gradientdescent}}
}
