% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{recurrentneuralnetwork_data}
\alias{recurrentneuralnetwork_data}
\title{Generate Recurrent Neural Network Data}
\usage{
recurrentneuralnetwork_data(n = 50000)
}
\description{
Generate Recurrent Neural Network Data
}
