% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{recurrentneuralnetwork_training}
\alias{recurrentneuralnetwork_training}
\title{Recurrent Neural Network Training}
\usage{
recurrentneuralnetwork_training(X1, X2, Y, hidden_dim = 10, alpha = 0.1)
}
\description{
Recurrent Neural Network Training
}
\seealso{
Other recurrent neural network: \code{\link{recurrentneuralnetwork_data}}

Other training: \code{\link{backpropagation_training}},
  \code{\link{deeplearning_training}},
  \code{\link{gradientdescent_training}}
}
