% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{gradientdescent}
\alias{gradientdescent}
\title{Gradient Descent}
\usage{
gradientdescent(X, Y)
}
\description{
Gradient Descent
}
\seealso{
Other training: \code{\link{backpropagation}},
  \code{\link{deeplearning}},
  \code{\link{recurrentneuralnetwork}}
}
