% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworks.R
\name{recurrentneuralnetwork_data}
\alias{recurrentneuralnetwork_data}
\title{Generate Recurrent Neural Network Data}
\usage{
recurrentneuralnetwork_data(n = 10000, binary_dim = 8)
}
\description{
Generate Recurrent Neural Network Data
}
\seealso{
Other data generators: \code{\link{backpropagation_data}},
  \code{\link{deeplearning_data}},
  \code{\link{gradientdescent_data}}
}
